Rename file for RedundancyPayFlow to match its class name
