+++ lib/smart_answer_flows/shared/redundancy_pay_flow.rb
